+++ ChestsAnywhere/release-notes.md
@@ -2,6 +2,7 @@
 
 # Release notes
 ## Upcoming release
+* Updated for the upcoming SMAPI 3.0.
 * Added button to reset chest options to their default.
 * Fixed issues saving chest options:
   * fixed form not saved if you only set the _(Automate) Put items in this chest first_ option;