+++ DataLayers/release-notes.md
@@ -1,5 +1,6 @@
 # Release notes
 ## Upcoming release
+* Updated for the upcoming SMAPI 3.0.
 * Improved translations. Thanks to Nanogamer7 (German)!
 
 ## 1.4.2